requirements.txt
asgiref==3.2.5
autopep8==1.5
Django==3.0.4
django-filter==2.2.0
django-rest-framework==0.1.0
djangorestframework==3.11.0
drf-action-serializer==1.0.5
pycodestyle==2.5.0
pytz==2019.3
sqlparse==0.3.1
